Redesign of model - without game

+++ Matchismo/Model/Card.h
@@ -11,11 +11,25 @@
 @interface Card : NSObject
 
 @property (strong, nonatomic) NSString *contents;
-
 @property (nonatomic, getter = isChosen) BOOL chosen;
 @property (nonatomic, getter = isMatched) BOOL matched;
 
-- (int)matchCount;
-- (int)match:(NSArray *)otherCards;
+/*
+ * matchCount:
+ * returns the number of cards
+ * needed to make a successful match
+ */
+- (NSUInteger)matchCount;
+
+/*
+ * match:
+ * returns the score of a match attempt
+ *   return value:
+ *     0: match failed
+ *     1: match successful
+ *   parameter
+ *     otherCards: is an array of cards to match with
+ */
+- (NSInteger)match:(NSArray *)otherCards;
 
 @end
